requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
telebot==0.0.5
urllib3==2.3.0
weatherapipython @ git+https://github.com/weatherapicom/python.git@05344f2d04cbf0b9513cb69b71f56a66a348e1be
